Moved completer to helpers folder
